agents/data_graph_multi_agent/requirements.txt
google-adk
python-dotenv
requests
beautifulsoup4
pydantic
uvicorn
fastapi
networkx
matplotlib
